mirage: Move `GET /api/settings` implementation into dedicated file
